+++ demo/src/main/java/com/SistemaVenta/demo/Repositorios/IBrandRepository.java
@@ -2,7 +2,7 @@ package com.SistemaVenta.demo.Repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.SistemaVenta.demo.Model.Brand;
 
-public interface IBrandRepository extends JpaRepository<Brand, Long> {
+public interface IBrandRepository extends JpaRepository<Brand, Integer> {
     // Additional query methods can be defined here if needed
 
 }
