+++ demo/src/main/java/com/SistemaVenta/demo/Repositorios/ICategoryRepository.java
@@ -2,7 +2,7 @@ package com.SistemaVenta.demo.Repositorios;
 import com.SistemaVenta.demo.Model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICategoryRepository extends JpaRepository<Category, Long> {
+public interface ICategoryRepository extends JpaRepository<Category, Integer> {
     // Aquí puedes definir métodos personalizados si es necesario   
 
 }
